% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_best_model.R
\name{geneseng_best_model}
\alias{geneseng_best_model}
\alias{geneseng_best_binary_log_model}
\alias{geneseng_best_linear_reg_model}
\title{Identify best features using a logistic/linear reg. with stepwise algorithm}
\usage{
geneseng_best_model(data, group, direction = "backward")

geneseng_best_binary_log_model(data, group, direction = "backward")

geneseng_best_linear_reg_model(data, group, direction = "backward")
}
\arguments{
\item{data}{provide a data.frame}

\item{group}{group by a categorical variable}

\item{direction}{direction of the stepwise algorithm}
}
\description{
Identify best features using a logistic/linear reg. with stepwise algorithm
}
