% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_summary_class_metrics.R
\name{geneseng_summary_class_metrics}
\alias{geneseng_summary_class_metrics}
\alias{geneseng_summary_binary_class_metrics}
\alias{geneseng_summary_multi_class_metrics}
\title{Summary of clinical performances}
\usage{
geneseng_summary_class_metrics(data, group, direction = c("auto", ">", "<"))

geneseng_summary_binary_class_metrics(
  data,
  group,
  direction = c("auto", ">", "<")
)

geneseng_summary_multi_class_metrics(
  data,
  group,
  direction = c("auto", ">", "<")
)
}
\arguments{
\item{data}{provide a data.frame}

\item{group}{group by a categorical variable}

\item{direction}{Define the positive case}
}
\description{
Evaluate biomarker individual performances using best classification metrics
}
\details{
`Logistic regression` was used for binary class problem. \cr
`Multinomial logistic regression` was used for multi class problem.
}
