% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_summary_corr.R
\name{geneseng_summary_corr}
\alias{geneseng_summary_corr}
\alias{continuous_to_continuous}
\alias{continuous_to_categ}
\alias{categ_to_categ}
\title{Evaluate correlation between all kind of variables}
\usage{
geneseng_summary_corr(data)

continuous_to_continuous(data)

continuous_to_categ(data)

categ_to_categ(data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
Evaluate correlation between all kind of variables
}
