% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_summary_stats.R
\name{geneseng_summary_stats}
\alias{geneseng_summary_stats}
\alias{make_summary}
\alias{make_summary_by_group}
\title{Summary of quantiative or categorical variables}
\usage{
geneseng_summary_stats(data, group = NULL)

make_summary(data)

make_summary_by_group(data, group)
}
\arguments{
\item{data}{provide a data.frame}

\item{group}{group by one variable}
}
\description{
Univariate or bivariate description of variables
}
