% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_summary_tests.R
\name{geneseng_summary_tests}
\alias{geneseng_summary_tests}
\alias{make_two_samples_tests}
\alias{make_two_or_more_samples_tests}
\title{Summary of statistical tests}
\usage{
geneseng_summary_tests(data, group)

make_two_samples_tests(data, group)

make_two_or_more_samples_tests(data, group)
}
\arguments{
\item{data}{provide a data.frame}

\item{group}{group by a categorical variable}
}
\description{
Statistical tests at 95% CI
}
