% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{spearmanCI}
\alias{spearmanCI}
\title{Compute Spearman coefficient and confident intervals at 95\%}
\usage{
spearmanCI(var1, var2)
}
\arguments{
\item{var1}{categorical variable}

\item{var2}{continuous variable}
}
\description{
Compute Spearman coefficient and confident intervals at 95\%
}
